# static/swagger.json
{
    "openapi": "3.0.3",
    "info":{ 
        "title":"Text and CSV Cleansing with Regex ",
        "version":"1.0.0",
        "description": "Clean the data",
        "contact": {
            "name": "binar",
            "email": "binar@gmail.com"
        },
        "license": {
            "name": "APACHE 2.0",
            "url": "htpps://www.apache.org/licenses/LICENSE-2.0"
        }
    },


    "servers":[
        {
            "description": "Text Cleansing RESTful API Server",
            "url": "/"
         }

    ],

    "externalDocs": {
        "description": "Github Code - Coba"
    },


    "paths":{
        "/":{
            "get":{
                "tags": ["Welcome Page"],
                "summary": "Welcome Page",
                "description": "Welcome to Cleansing Text API",
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            }
        },

        "/data":{
            "get":{
                "tags": ["Text Cleansing"],
                "summary": "Get all of data list",
                "description": "Thousand of data list",
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            },
            "post":{
                "tags": ["Text Cleansing"],
                "summary": "Post a text",
                "description": "Post a text file to clean",
                "requestBody":{
                    "required": true,
                    "content": {
                        "multipart/form-data":{
                            "schema":{
                                "type": "object",
                                "properties": {
                                    "text":{
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            }
        },



        "/data/csv":{
            "post":{
                "tags": ["CSV Cleansing"],
                "summary": "Post a csv file",
                "description": "Post a csv file to clean",
                "requestBody":{
                    "required": true,
                    "content": {
                        "multipart/form-data":{
                            "schema":{
                                "type": "object",
                                "properties": {
                                    "file":{
                                        "type": "string",
                                        "format": "base64"
                                    }
                                }
   
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            }
        }

    }
}